chore(vault): sync lockfile for   knip, drop orphaned stepDescriptions
